moved parser to mail_processing folder
